src/backend/requirements.txt
fastapi==0.115.6
pydantic==1.10.19
fastapi-jwt-auth==0.5.0
yfinance==0.2.20
pandas==1.5.3
pytz==2023.3
passlib==1.7.4
pymongo==4.3.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pymongo==4.6.0
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2
tensorflow==2.15.0
scikit-learn==1.3.2
sse-starlette==1.6.5
python-multipart==0.0.6
fastapi
uvicorn
python-dotenv
pymongo
passlib[bcrypt]
python-jose
pydantic
yfinance
pandas
numpy
talib-binary
tensorflow
scikit-learn
sse-starlette
